Generalizes the use of dataloaders
Dataloaders can be objects that either return two or three arguments. A third argument that is being used with monkey data are "valid responses", meaning that the neuron in that session has actually seen the input image. The "Valid Responses" argument is now being referred to as "weights", generalizing its use.
Also changes the functionality of the "cuda" input argument, so that devices can be user-specified.